app/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
